Load testing development (#1)
* Add load-test definitions in grafana k6

* Update the neo4j test definitions ...

Replace the hardcoded username and password with an
environment variable. Corrects the query syntax for the MATCH
cypher query

* Add the elasticsearch load tests

* Add sampling module to simplify testing logic

* Test modularization ...

Adds a configuration file for different environments (for running
locally and for running on su08). Provides different scenarios
to combine the basic smoke test with the load testing in one test file

* Expand testcase options and scenarios

* Expand the README documentation

* Update dockerfile and docker-compose along with docs